Drop unnecessary int cast in File::getTotalScore
PHPStan 2.2.6 already infers array_sum over array<string, int> as int,
so the cast is redundant.

+++ php/Parser/File.php
@@ -21,7 +21,7 @@ final class File extends Parser
 
     protected function getTotalScore(): int
     {
-        return (int) array_sum(array_map(array_sum(...), $this->model));
+        return array_sum(array_map(array_sum(...), $this->model));
     }
 
     protected function getScore(string $featureKey, string $sequence): int
